% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorling-params.R
\name{dorling_from_sp}
\alias{dorling_from_sp}
\title{Compute center coordinates, circle radii and Dorling polygons from input polygons and values}
\usage{
dorling_from_sp(polys, value, n_rescales = 50, n_iter = 50, tol = 1000,
  quiet = TRUE)
}
\arguments{
\item{polys}{\code{SpatialPolygonsDataFrame}}

\item{value}{values that correspond to the ordered polygons in \code{polys}}

\item{n_rescales, n_iter, tol}{tuning parameters for computing non-overlapping
circle polygon radius values. Tweak only if the defaults do not work well.}

\item{quit}{if \code{TRUE} (default) do not show the scaling iterations}
}
\value{
a \code{list} with \code{xy} coordinate matrix, \code{radius} vector, and disc \code{SpatialPolygons}
}
\description{
Pass in an equal area (not long-lat) projected \code{SpatialPolygonsDataFrame} and get
back what you need to plot a Doriling cargogram.
}
\note{
There is a significant limitation in that the polygons have to all be touching
(i.e. this is why the README has an example w/o Alaska and Hawaii). If
neighbors were computed separately or allowed to be passed in, this
could be made more generic. PRs & Issues welcome.
}
\author{
Mathieu Rajerison \url{https://datagistips.hypotheses.org/}
}
