% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdorling-package.R
\docType{package}
\name{spdorling}
\alias{spdorling}
\alias{spdorling-package}
\title{Create Dorling Cartograms from Contiguous Region Shapefiles}
\description{
Create Dorling Cartograms from Contiguous Region Shapefiles
}
\author{
Bob Rudis (bob@rud.is)

Mathieu Rajerison \url{https://datagistips.hypotheses.org/}
}
